impl create tweet

+++ db/migrate/20181228001800_create_tweets.rb
@@ -2,7 +2,7 @@ class CreateTweets < ActiveRecord::Migration[5.2]
   def change
     create_table :tweets do |t|
       t.references :user, null: false, foreign_key: true
-      t.integer :tweet_id, null: false
+      t.bigint :tweet_id, null: false
 
       t.timestamps
     end
